+++ src/com/company/Star.java
@@ -8,7 +8,7 @@ public class Star extends GravBod{
 
     private int mass;
 
-    public Star(int x, int y, int radius, Color color, int mass) {
+    public Star(double x, double y, int radius, Color color, int mass) {
         super(x, y, radius, color);
         this.mass = mass;
     }
